New and improved API for modeldict.
Now, any implementation of a "ModelDict" is actually a subclass of a "PersistentDict" object.  The subclass is responsible for implementing a set of methods to read, write and delete data to its persistent store, plus report back a comparable object for the "last_updated" value.  PersistentDict simply proxies all methods to the underlying dict for storage, but makes sure writes and deletes go to persistent storage.  The cached data in process for the dict is refreshed from persistent storage once the storage changes, noticeable by asking the class for its "last_updated" value.  The implementation of PersistentDict checks the last_updated before every method, so it must be fast.

This works well for Redis, where we can store the configuration as a hash and keep the last_updated as an int and atomically increment it with the redis INCR command.

For other implementation, like the Django model one, they will need a relational DB to keep the settings config in and a faster data store, like memcache to keep the last_updated value.